stricter test conditions
